These are some of the projects I did during the year 2023-2024.
This project involves creating Verilog code for three different components: a 2x1 Quad Mux, a BCD to 7-segment driver, and a 2-to-4 decoder. The tasks include writing the Verilog descriptions for each module, connecting them within the Quartus tool to form a functional block diagram, and simulating the components to verify their functionality. The project requires you to document your work with the code, Quartus block diagrams, and screenshots from the simulations. The final submission will include all the relevant files, diagrams, and simulation results.