Digital Circuits Implementation using Breadboard , Sequential Logic Circuits using Breadboard and IC’s
In this experiment, electronic circuits were designed using ICs on a breadboard. The focus was on utilizing ICs such as 74xx to implement circuits like a decoder and a BCD to seven-segment display decoder. Circuits were built using logical gates such as AND and NOT, connecting them to represent inputs and produce the required outputs. A 7490 counter was used to count from 0 to 9, displaying the results on a seven-segment display. Proper power connections were ensured, and the circuits were implemented according to the diagrams for each step, with tests conducted to verify the output accuracy.